fix issue #82

+++ application/modules/dashboard/views/case_field.php
@@ -43,12 +43,14 @@ if($posts){
                       <?php
                       $group_name = null;
                       for($ix=0;$ix<count($user_list);$ix++){
-                                 //if(IsRoleFriendlyNameExist($user_list[$i]->role_detail, 'Social Stream_Current_Resolve_Case')){
+                                 if(IsRoleFriendlyNameExist($user_list[$i]->role_detail, 'Social Stream_Current_Resolve_Case')){
                                             if($user_list[$ix]->group_name!=$group_name){
                                                        echo '<optgroup label="'.$user_list[$ix]->group_name.'"></optgroup>';           
                                             }
-                                            echo '<option value="'.$post_id.'-'.$user_list[$ix]->user_id.'">'.$user_list[$ix]->full_name.'</option>';
-                                 //}
+                                            if( $this->session->userdata('user_id') != $user_list[$ix]->user_id){
+                                                       echo '<option value="'.$post_id.'-'.$user_list[$ix]->user_id.'">'.$user_list[$ix]->full_name.'</option>';                                 
+                                            }
+                                 }
                                  $group_name = $user_list[$ix]->group_name;           
                       }
                       ?>
